addition of google login, and changes into others entities

+++ src/files/files.service.ts
@@ -5,19 +5,35 @@ import { ILike, Repository } from 'typeorm';
 import { FileEntity } from 'src/entities/file.entity';
 import { CreateFileDto } from './domain/dtos/create-file.dto';
 import { UpdateFileDto } from './domain/dtos/update-file.dto';
+import { AzureBlobService } from './integration/azure.service';
+import { v4 as uuidv4 } from 'uuid';
+import { extname } from 'path';
 
 @Injectable()
 export class FilesService {
   constructor(
     @InjectRepository(FileEntity)
     private fileRepository: Repository<FileEntity>,
+    private readonly azureService: AzureBlobService,
   ) {}
 
-  async createFile(createFileDto: CreateFileDto): Promise<FileEntity> {
+  async createFile(file: Express.Multer.File): Promise<FileEntity> {
     try {
-      const hero = this.fileRepository.create(createFileDto);
+      const fileExtension: string = extname(file.originalname);
+      const blobName: string = `${uuidv4()}${fileExtension}`;
 
-      return this.fileRepository.save(hero);
+      await this.azureService.uploadBlobToContainer(blobName, file.buffer);
+
+      const url = await this.azureService.getFileUrl(blobName);
+
+      const newFile: CreateFileDto = { filename: file.originalname, url: url };
+
+      const hero = this.fileRepository.create({
+        blobName: blobName,
+        ...newFile,
+      });
+
+      return await this.fileRepository.save(hero);
     } catch (error) {
       throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
@@ -74,13 +90,23 @@ export class FilesService {
     }
   }
 
-  async deleteUser(id: string): Promise<boolean> {
+  async deleteFile(id: string): Promise<boolean> {
     try {
-      const enterprise = await this.fileRepository.delete({ id: id });
-      const checkEnterprise = await this.fileRepository.exist({
-        where: { id: id },
+      const enterpriseGet = await this.fileRepository.findOne({
+        where: { id },
       });
-      return !checkEnterprise;
+
+      if (enterpriseGet) {
+        await this.fileRepository.delete({ id: id });
+
+        await this.azureService.deleteBlob(enterpriseGet.blobName);
+
+        const checkEnterprise = await this.fileRepository.exist({
+          where: { id: id },
+        });
+
+        return checkEnterprise;
+      }
     } catch (error) {
       throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
